Search interface appeared! You can find any information you want using new Search().
It is pre-release version of Search interface - it is only for testing purposes, it is do not have real pages search, only index.

Here what we can do:
This search is similar to native language. Just simply write something like that:
$search = new Search();
$pages = $search->from(0)->to(25)->byType("tshirt")->where("main","size")->isEqualTo("M")->find();
And that's it! You will get maximum 25 "tshirts" pages only with "M" "size".

Or, for example you want to get full pages URL-index. No problem, use this:
$search = new Search();
$pages = $search->findAsList();